==> src/notebooks/entrega1/1_landing_zone.ipynb <==
# **Data Collectors and Landing Zone**

### *Required packages:*

 !pip install opendatasets
 !pip install kaggle
 !pip install openpyxl
 !pip install pyarrow

import numpy as np
import pandas as pd
import requests
import os
import opendatasets as od
import csv
import openpyxl
import pyarrow

### Kaggle Configuration - Drive

(Requereix del fitxer `kaggle_fitxer.json`)

!mkdir ~/.kaggle

!wget https://raw.githubusercontent.com/OscarMoliina/betterlifebetterhealth/main/src/utils/kaggle_user_config.json

!cp ./kaggle_user_config.json ~/.kaggle/kaggle.json

!cd ~/.kaggle && ls

### Downloading datasets

#CREAR DIRECTORIOS SI NO EXISTEN:

# Nombre del directorio
nombre_directorio = "data"

# Verificar si el directorio existe
if not os.path.exists(nombre_directorio):
    # Crear el directorio
    os.makedirs(nombre_directorio)
    print(f"Directorio '{nombre_directorio}' creado exitosamente.")
else:
    print(f"El directorio '{nombre_directorio}' ya existe.")

%cd data

nombre_directorio = "csv"

# Verificar si el directorio existe
if not os.path.exists(nombre_directorio):
    # Crear el directorio
    os.makedirs(nombre_directorio)
    print(f"Directorio '{nombre_directorio}' creado exitosamente.")
else:
    print(f"El directorio '{nombre_directorio}' ya existe.")

nombre_directorio = "parquet"

# Verificar si el directorio existe
if not os.path.exists(nombre_directorio):
    # Crear el directorio
    os.makedirs(nombre_directorio)
    print(f"Directorio '{nombre_directorio}' creado exitosamente.")
else:
    print(f"El directorio '{nombre_directorio}' ya existe.")

%cd ..

##### UNECE

# URL del dataset UNECE
unece_db = pd.read_json('https://query.data.world/s/qd6kjqguuvb573mffbpfysgpudi2pv?dws=00000')
unece_db.to_csv('unece_db.csv', index=False)
!mv unece_db.csv data/csv/unece_db.csv

##### Kaggle - Mental Health

# API Kaggle Mental Health
!kaggle datasets download  -d 'thedevastator/uncover-global-trends-in-mental-health-disorder'

#!sudo apt-get install unzip
!unzip uncover-global-trends-in-mental-health-disorder.zip -d data/csv
!rm uncover-global-trends-in-mental-health-disorder.zip

##### USA Census

# API United States Census
variables = "NAME","YR","AREA_KM2","CBR","CDR","DEATHS","E0","GRR","IMR","MEDAGE","MR0_4","POP_DENS","genc standard countries and areas"
url_base = "https://api.census.gov/data/timeseries/idb/5year?get=NAME,YR,AREA_KM2,CBR,CDR,DEATHS,E0,GRR,IMR,MEDAGE,MR0_4,POP_DENS&for=genc%20standard%20countries%20and%20areas:*".format(variables)

# Llista per guardar les dades
data = []

# Fem una sol·licitud a l'API per agafar dades des de l'any 1990 fins al 2019
for year in range(1990, 2019):
    url = "{}&time={}".format(url_base, year)
    response = requests.get(url)
    if response.status_code == 200:
        data.extend(response.json()[1:])

# Guardar los datos en un archivo CSV
csv_filename = "dades_demogr_census.csv"
with open(csv_filename, "w", newline="") as csv_file:
    csv_writer = csv.writer(csv_file)
    csv_writer.writerow(["NAME","YR","AREA_KM2","CBR","CDR","DEATHS","E0","GRR","IMR","MEDAGE","MR0_4","POP_DENS","genc standard countries and areas", "code"])
    csv_writer.writerows(data)

print("Les dades s'han guardat a", csv_filename)

!mv dades_demogr_census.csv data/csv/dades_demogr_census.csv


#### Rename

!mv data/csv/dades_demogr_census.csv data/csv/demography.csv
!mv data/csv/unece_db.csv data/csv/society.csv
os.rename("data/csv/Mental health Depression disorder Data.csv", "data/csv/mental_health.csv")


#### Storing files in a suitable format (Parquet)


# Unece_db
unece_db = pd.read_csv('data/csv/society.csv')
unece_db.to_parquet('data/parquet/society.parquet')

# Kaggle db
kaggle_db = pd.read_csv('data/csv/mental_health.csv', decimal=',')
kaggle_db.to_parquet('data/parquet/mental_health.parquet')

# United State Census db
us_census_db = pd.read_csv('data/csv/demography.csv')
us_census_db.to_parquet('data/parquet/demography.parquet')

#### Guardar parquets:

# Clone the repository
!git clone https://github.com/OscarMoliina/betterlifebetterhealth.git
%cd betterlifebetterhealth/data/parquet

!rm -rf *
!cp -R /content/data/parquet/* .

!git config --global credential.helper store

#CODI A LA DOCUMENTACIO NO POSAT AQUÍ PER PRIVACITAT DE USUARI
