==> src/notebooks/entrega1/5_model_pipeline.ipynb <==
#!pip3 install scikit-learn
!pip3 install xgboost
!pip3 install pyspark
!pip3 install duckdb

import pyspark
from pyspark.sql import SparkSession

import pandas as pd

import duckdb

from sklearn.model_selection import train_test_split
from sklearn.svm import SVR
from sklearn.metrics import mean_squared_error
from sklearn.preprocessing import StandardScaler

import joblib

!git clone https://github.com/OscarMoliina/betterlifebetterhealth.git

spark = SparkSession.builder \
    .appName("Preprocessing") \
    .config("spark.jars", "/content/betterlifebetterhealth/src/utils/duckdb.jar") \
    .getOrCreate()

data = spark.read \
    .format("jdbc") \
    .option("url", "jdbc:duckdb:/content/betterlifebetterhealth/data/db/exploitation_zone.db") \
    .option("driver", "org.duckdb.DuckDBDriver") \
    .option("dbtable", "join_table") \
    .load()

# Comparativa Dataset Simple vs. Dataset Mitxe

df = data.toPandas()

y = df['Depression (%)']
X1 = df[['Schizophrenia (%)', 'Bipolar disorder (%)', 'Eating disorders (%)',
         'Anxiety disorders (%)', 'Drug use disorders (%)', 'Alcohol use disorders (%)']]
X2 = df.drop(['Country', 'Year', 'Depression (%)'], axis=1)

X1_train, X1_test, y_train, y_test = train_test_split(X1, y, test_size=0.2, random_state=2003)
X2_train, X2_test = train_test_split(X2, test_size=0.2, random_state=2003)

scaler = StandardScaler()
X1_train_scaled = scaler.fit_transform(X1_train)
X1_test_scaled = scaler.transform(X1_test)
X2_train_scaled = scaler.fit_transform(X2_train)
X2_test_scaled = scaler.transform(X2_test)

svm_model1 = SVR()
svm_model2 = SVR()
svm_model1.fit(X1_train_scaled, y_train)
svm_model2.fit(X2_train_scaled, y_train)

# joblib.dump(scaler, 'scaler1.pkl')
# joblib.dump(svm_model1, 'svm_model1.pkl')
# joblib.dump(svm_model2, 'svm_model2.pkl')

y_pred1 = svm_model1.predict(X1_test_scaled)
y_pred2 = svm_model2.predict(X2_test_scaled)
mse1 = mean_squared_error(y_test, y_pred1)
mse2 = mean_squared_error(y_test, y_pred2)

print("MSE para el modelo con solo datos de enfermedades:", mse1)
print("MSE para el modelo con todas las columnas:", mse2)


# Model Pipeline

import joblib
from pyspark.sql import DataFrame
from sklearn.metrics import mean_squared_error
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.svm import SVR
from xgboost import XGBRegressor

class ModelPipeline:
    def __init__(self, data: DataFrame, objective:str, model_type='SVM', model=None, scaler=None) -> None:
        self.data = data
        self.objective = objective
        self.model_type = model_type
        self.model = model
        self.scaler = scaler if scaler is not None else StandardScaler()

    def __train_model(self):
        df = self.data.toPandas()
        y = df[self.objective]
        X = df.drop(['Country', 'Year', self.objective], axis=1)
        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=2003)

        X_train_scaled = self.scaler.fit_transform(X_train)
        X_test_scaled = self.scaler.transform(X_test)

        if self.model_type == 'SVM':
            self.model = SVR()
        elif self.model_type == 'RF':
            self.model = RandomForestRegressor()
        elif self.model_type == 'XGB':
            self.model = XGBRegressor(objective='reg:squarederror')

        self.model.fit(X_train_scaled, y_train)
        y_pred = self.model.predict(X_test_scaled)
        mse = mean_squared_error(y_test, y_pred)

        joblib.dump(self.model, f'{self.model_type}_model.pkl')
        joblib.dump(self.scaler, 'scaler.pkl')

        return self.model, self.scaler, None, mse

    def __make_predictions(self):
        df = self.data.toPandas()
        X = df.drop(['Country', 'Year', self.objective], axis=1)
        X_scaled = self.scaler.transform(X)

        y_pred = self.model.predict(X_scaled)

        return self.model, self.scaler, y_pred, None

    def predict(self):
        if self.model is None or isinstance(self.model, str) and not self.model.endswith('.pkl'):
            return self.__train_model()
        else:
            self.model = joblib.load(self.model) if isinstance(self.model, str) else self.model
            self.scaler = joblib.load(self.scaler) if isinstance(self.scaler, str) else self.scaler
            return self.__make_predictions()

# Exemples d'Ús

pipline = ModelPipeline(data=data, objective='Depression (%)', model_type='XGB', scaler=None, model=None)
model, scaler, predictions, mse = pipline.predict()

print(model)
print(scaler)
print(predictions)
print(mse)

pipline = ModelPipeline(data=data, objective='Depression (%)',model_type='SVM')
model, scaler, predictions, mse = pipline.predict()

print(model)
print(scaler)
print(predictions)
print(mse)

pipline = ModelPipeline(data=data, objective='Depression (%)',model_type='RF')
model, scaler, predictions, mse = pipline.predict()

print(model)
print(scaler)
print(predictions)
print(mse)
